refactor: update cjsToEsm.js to use ES6 modules
