+++ src/modules/cjsToEsm.mjs
@@ -1,17 +1,24 @@
-const path = require('path');
-const { release, version } = require('os');
-const { createServer: createServerHttp } = require('http');
-require('./files/c');
+import * as path from 'node:path';
+import { release, version } from 'node:os';
+import { createServer as createServerHttp } from 'node:http';
+import { fileURLToPath } from 'node:url';
+import { readFile } from 'fs/promises';
+import './files/c.mjs';
 
 const random = Math.random();
 
 let unknownObject;
 
-if (random > 0.5) {
-    unknownObject = require('./files/a.json');
-} else {
-    unknownObject = require('./files/b.json');
-}
+const jsonFilePath = random > 0.5 ? './files/a.json' : './files/b.json';
+
+unknownObject = JSON.parse(
+  await readFile(
+    new URL(jsonFilePath, import.meta.url)
+  )
+);
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 console.log(`Release ${release()}`);
 console.log(`Version ${version()}`);
@@ -33,8 +40,7 @@ myServer.listen(PORT, () => {
     console.log('To terminate it, use Ctrl+C combination');
 });
 
-module.exports = {
+export {
     unknownObject,
     myServer,
 };
-
