+++ src/modules/testMjs.js
@@ -1,0 +1,3 @@
+import { unknownObject } from "./cjsToEsm.mjs";
+
+console.log('Print unknownObject in testMjs:', unknownObject);
